refactor: rename files and switch to pytest
