+++ src/suricatals/langserver.py
@@ -31,7 +31,7 @@ from importlib.metadata import version
 import queue
 
 from suricatals.parse_signatures import SuricataFile
-from suricatals.tests_rules import TestRules
+from suricatals.signature_validator import TestRules
 from suricatals.suri_cmd import SuriCmd
 from suricatals.tokenize_sig import SuricataSemanticTokenParser
