+++ src/suricatals/worker_pool.py
@@ -23,7 +23,7 @@ import queue
 from typing import Dict, Optional, Tuple, Any
 
 from suricatals.parse_signatures import SuricataFile
-from suricatals.tests_rules import TestRules
+from suricatals.signature_validator import TestRules
 
 log = logging.getLogger(__name__)
